Oanda: Prune unnecessary API request

+++ cpp/sandbox/src/alpaca/tao_bot/set_execution_price.cpp
@@ -28,7 +28,7 @@ void Alpaca::TaoBot::set_execution_price(order_t *order) {
 
   const std::string execution_price_string = order_json["filled_avg_price"];
 
-  order->execution_price = std::stod(execution_price_string);
+  order->execution_price = (double)std::stod(execution_price_string);
 }
 
 #endif
